// apps/administration/src/core/permission/application/dto/out/permission-deleted-response-dto.ts
/* ============================================
   administration/src/core/permission/application/dto/out/permission-deleted-response-dto.ts
   ============================================ */

export interface PermissionDeletedResponseDto {
  id_permiso: number;
  message: string;
  deletedAt: Date;
}